Add auth router to main app
