+++ tsconfig.json
@@ -31,8 +31,8 @@
       // "baseUrl": "./",                                  /* Specify the base directory to resolve non-relative module names. */
       // "paths": {},                                      /* Specify a set of entries that re-map imports to additional lookup locations. */
       // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
-      "typeRoots": ["./types"],                                  /* Specify multiple folders that act like './node_modules/@types'. */
-      "types": ["./types", "./node_modules/@types"],                                      /* Specify type package names to be included without being referenced in a source file. */
+      // "typeRoots": ["./types"],                                  /* Specify multiple folders that act like './node_modules/@types'. */
+      // "types": ["./types", "./node_modules/@types"],                                      /* Specify type package names to be included without being referenced in a source file. */
       // "allowUmdGlobalAccess": true,                     /* Allow accessing UMD globals from modules. */
       // "moduleSuffixes": [],                             /* List of file name suffixes to search when resolving a module. */
       "resolveJsonModule": true,                        /* Enable importing .json files. */
