--- types/auth/index.d.ts
@@ -1,31 +1,0 @@
-import { RequestHandler, Router } from "express";
-
-type AuthPack = {
-  router: Router;
-  authMiddleware: RequestHandler;
-};
-
-export const AuthSchemes = {
-  JWT: "jwt",
-  HTTP: "http",
-};
-
-export type AuthSchemes = typeof AuthSchemes[keyof typeof AuthSchemes];
-
-export interface HttpOptions {
-  expire?: number;
-  secret?: string;
-}
-
-export interface JwtOptions {
-  accessExpire?: number;
-  refreshExpire?: number;
-  secret?: string;
-}
-
-export type JWTPayload = {
-  id: string;
-  iat: number | undefined;
-  exp: number | undefined;
-  type?: stgring;
-};
